requirements.txt
telethon
ffmpeg-python
imageio-ffmpeg
mutagen
pillow